tombstoning this effort. needs to be redone from scratch by the recipe in the readme. its a tug of war between old versions that can no longer install or work against the Internet repositories as they are today, and new versions that don't like elements of how the build (which was coded to old versions) is.
work within that opening which these two boundaries make.

the useful files to keep from here are the readme and verify.py. maybe also Dockerfile which has the underlying OS installs which are required, though that's the easy part to redo and maybe better let AI do that part from scratch to start clean.