include .git in the image ― the image will have your current mediapipe project with its included .git in it rather than the project without its .git. this enables building in the image as its modified build uses git for __version__ naming which is required in modern pip. but also this add traceability anyway.
the image is now built with all mediapipe v0.10.13 build dependencies because it runs the mediapipe build as part of its creation.

this makes the image building slower by a few mintues (the build duration) and it makes it a little bigger in size, but it future proofs having all those bazel and pip dependencies some of which will get older and disappear from Internet server sooner than others.