port from branch liberation the c++ main running the hands pipeline directly, not the example c++ running it which does not expose the pipeline outputs directly by wrapping around that pipeline.
unlike previous commit comment suggested, by this commit we build the necessary pipeline output data protobuf message type to enable the above.

this does not change the hands pipeline's source code nor bazel building code.

it only adds a runner of the inner pipeline which is the one we actually use from python.

+ include