more informed way of making it build opencv from source, which also thoroughly inline documents where it configures whether to use the system installed (which it does, prior to this commit, regardless the -I flag, as it is configured to do so by default, -I probably only explicitly needed for the python package).
opencv 4.10 source is downloaded and built.

opencv compilation is of course, long.

Run from the C API lib via rust (but not when running the C++ mains) it still dynamically loads the system libprotobuf upon initialization:

calling init: /lib/x86_64-linux-gnu/libprotobuf.so.32

and the dnn module of opencv which is known to bring in protobuf:

calling init: /lib/x86_64-linux-gnu/libopencv_dnn.so.406

then fails same as before.