after more work on avoiding the OpenCV crate bring in the system OpenCV, making it use our same bazel built one, the protobuf issue is solved now in this commit.
this also required tethering some not only protobuf dependencies as always linked, for our lib build target, because alwayslink only cascades through transitive dependencies if all of tham have alwayslink, not if any one dep in the chain does not (it's enough for one dep in the chain not to have alwasylink=1, to make its children's symbols not linked, even if they themselves have alwayslinke=1).

so tethered the necessary ones by brute force, becasue on the current branch bazel query can't be used for a more minimal tethering by examining the existing paths systematically, which would have potentially (not surely) enabled just adding some alwayslink=1 in existing definitions instead of duplicating deps in a higher node like we did).