towards extracting a non framework Context object reliant inference calculator:
both inference models now use code that's maximally cleaned from context use ― only using context for input and output wiring ―

previously there was a reliance on the calling calculator class which made the core's work entangled to rely on the calculator being a calculator ― now this entanglement has been removed and so ― the core no longer has that dependency.

techincally this disentanglement was accomplished by:

(A) mimicking how an original inference calculator gets its input (with help from the dummy calculator blueprint we had) which removed the need to to access

(B) switching our custom inference class DetectionInferenceCalculator to inherit the Node base class more directly thus making it get all the features it needs built in.

(C) the combination of these untangled it from needing to fiddle the io_mapper_ member which entangled the extracted core to rely on the calculator using it for no good reason but probably being just an aretefact of the initial porting process somehow.