core the AssociationNormRectCalculator calculator (AssociationCalculator) such that it performs the simple greedy merge and filter operation without the over-generalized logic and logic pieces which it would never reach even in the v0.10.13 code revision.
inline documentation of its hyper-generalized logic is therefore removed by this commit.

verified against multiple videos, with num_hands increased up to 3, and with confirmation of no hand ids being passed to this calculator even by debug logging in the baseline branch.

the documentation of what this made-lean node does is left in the pbtxt file.