1. wrapping up a 1.5 days battle around regressions in how the C api is linked into rust, which started in noticing that rust fails to link OpenCV.
2. bringing the C api use from the C++ main also back in working condition while troubleshooting the differences.

3. this push overlapped with making the current mediapipe version build too, which is on a separate repo "nui-ai/mediapipe-current" but may have inspired some small changes included here as well.

a cascade of fixes, including an absl api update which somehow became necessary ensued and is included in this commit, mostly all of which inline self-documented.

notably some of the changes were about the base directory from where things are loaded, and the rust code now changes its own working directory so that it (for now) avoids passing a root directory to our mediapipe wrapper for the latter's loading of model files.

issues opened during the work:

https://github.com/twistedfall/opencv-rust/issues/697
https://github.com/google-ai-edge/mediapipe/issues/6132
https://github.com/google-ai-edge/mediapipe/issues/6130