1. make the expected number of hands a required (not default-valued) input argument in all paths instantiating the pipeline and the no-pipeline, up to program arguments of all mains, so that this meta parameter is always explicit ― to avoid future workflows which assume that number wrongly or do not log it.
all run configurations accordingly adapted to explicitly provide it.

the above was slightly vexing when revisiting our pipeline runner wrapper, whose documentation was made now much more correct to avoid the same.

3. unrelated to the above: fix the C API run configuration to use the new correct pbtxt file name that was renamed few commits back.

ultimately we'll have to write metaparams of inference data being persisted in or next to their data files ― which we don't even do for the max number of hands argument which has been the topic of this commit ― this commit is only the first out of two steps for inference data safety in terms of reproducibility in liu of this number being a variable for pipeline and no-pipeline runs alike in these baseline commit levels of full liberation where we are now.