temp commit showing how working to rename our python namespace is a whole week project .... the proto files are one big hurdle as they cannot be renamed are sprinkled around and found by a longish glob list, cannot be renamed while being generated, and reference each other by name. the first two traits are those which make it horrible.
whereas not firmly knowing which ones the mediapipe api really needs for python makes any process x10 more winding ... that's because:

this is not like a bazel build ― you do not have a written definition of which python files require which ones!

it's doable for sure with a half a week's time ― but it is not a project goal to provide co-tenancy for a python project!!! hence this start will be reverted on the next commit.